phase2/requirements.txt
torch
torchvision
numpy
scipy
scikit-learn
matplotlib
pyyaml

